+++ docs/html/search/files_4.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['string_5fbuffer_2eh_204',['string_buffer.h',['../string__buffer_8h.html',1,'']]]
+  ['string_5fbuffer_2eh_245',['string_buffer.h',['../string__buffer_8h.html',1,'']]]
 ];