+++ src/ace_time_c/string_buffer.h
@@ -1,5 +1,7 @@
-#ifndef ACE_TIME_C_STRING_BUFFER_H
-#define ACE_TIME_C_STRING_BUFFER_H
+/*
+ * MIT License
+ * Copyright (c) 2022 Brian T. Park
+ */
 
 /**
  * @file string_buffer.h
@@ -7,8 +9,15 @@
  * A simple set of integer to string conversion functions.
  */
 
+#ifndef ACE_TIME_C_STRING_BUFFER_H
+#define ACE_TIME_C_STRING_BUFFER_H
+
 #include <stdint.h>
 #include <stddef.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 /** An array of characters that knows its capacity and its current usage size */
 typedef struct AtcStringBuffer {
@@ -21,10 +30,13 @@ typedef struct AtcStringBuffer {
 } AtcStringBuffer;
 
 /** Initialize the AtcStringBuffer with its char pointer and capacity. */
-void atc_print_init(AtcStringBuffer *sb, char *p, uint8_t capacity);
+void atc_buf_init(AtcStringBuffer *sb, char *p, uint8_t capacity);
+
+/** Reset the buffer to zero length. */
+void atc_buf_reset(AtcStringBuffer *sb);
 
 /** Finalize the AtcStringBuffer by terminating the string with a NUL. */
-void atc_print_end(AtcStringBuffer *sb);
+void atc_buf_close(AtcStringBuffer *sb);
 
 /** Print one char into the AtcStringBuffer. */
 void atc_print_char(AtcStringBuffer *sb, char c);
@@ -43,5 +55,9 @@ void atc_print_uint16_pad4(AtcStringBuffer *sb, uint16_t n);
 
 /** Print the signed integer. A negative number will print a leading "-". */
 void atc_print_int16(AtcStringBuffer *sb, int16_t n);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif