+++ src/ace_time_c/zone_registrar.h
@@ -6,8 +6,8 @@
 /**
  * @file zone_registrar.h
  *
- * Functions to find the AtcZoneInfo from the human readable zone name,
- * or the 32-bit zone ID identifier.
+ * Functions that search for specific time zones by name or by ID over a given
+ * set of zones defined by a zone registry.
  */
 
 #ifndef ACE_TIME_C_ZONE_REGISTRAR_H
@@ -17,6 +17,23 @@
 #include <stdbool.h>
 #include "zone_info.h"
 #include "common.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/** Data structure used by the registrar to manage a given zone registry. */
+typedef struct AtcZoneRegistrar {
+  const AtcZoneInfo * const * registry;
+  uint16_t size;
+  bool is_sorted;
+} AtcZoneRegistrar;
+
+/**Initialize the given registrar data structure with the given registry. */
+void atc_registrar_init(
+    AtcZoneRegistrar *registrar,
+    const AtcZoneInfo * const * registry,
+    uint16_t size);
 
 /** Determine if the registry is sorted by zone id. */
 bool atc_registrar_is_registry_sorted(
@@ -28,19 +45,19 @@ bool atc_registrar_is_registry_sorted(
  * Return NULL if not found.
  */
 const AtcZoneInfo *atc_registrar_find_by_name(
-    const AtcZoneInfo * const * registry,
-    uint16_t size,
-    const char *name,
-    bool is_sorted);
+    const AtcZoneRegistrar *registrar,
+    const char *name);
 
 /**
  * Search the zone registry for the zone 'id'.
  * Return NULL if not found.
  */
 const AtcZoneInfo *atc_registrar_find_by_id(
-    const AtcZoneInfo * const * registry,
-    uint16_t size,
-    uint32_t zone_id,
-    bool is_sorted);
+    const AtcZoneRegistrar *registrar,
+    uint32_t zone_id);
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif