+++ tests/zonedb_test.c
@@ -6,8 +6,8 @@ ACU_TEST(test_zonedb_sizes)
   // These numbers are correct for TZDB 2022a
   ACU_ASSERT(sizeof(kAtcZoneRegistry) / sizeof(AtcZoneInfo*) == 351);
   ACU_ASSERT(sizeof(kAtcZoneAndLinkRegistry) / sizeof(AtcZoneInfo*)
-      == 595);
-  ACU_ASSERT(sizeof(kAtcLinkRegistry) / sizeof(AtcLinkEntry) == 244);
+      == 596);
+  ACU_ASSERT(sizeof(kAtcLinkRegistry) / sizeof(AtcLinkEntry) == 245);
 }
 
 //---------------------------------------------------------------------------